+++ src/Nox.Reference.GetData/MacAddresses/MacAddressInfo.cs
@@ -1,0 +1,13 @@
+﻿using Nox.Reference.Abstractions.MacAddresses;
+
+internal record MacAddressInfo : IMacAddressInfo
+{
+    public MacAddressInfo(string address, string vendor)
+    {
+        Address = address;
+        Vendor = vendor;
+    }
+
+    public string Address { get; }
+    public string Vendor { get; }
+}
