+++ src/Nox.Reference.MacAddresses/Nox.Reference.MacAddresses.csproj
@@ -32,6 +32,5 @@
 
 	<ItemGroup>
 	  <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
-	  <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 	</ItemGroup>
 </Project>
